Fudge renewal process
This is entirely for dev purposes. It selects all of the routes (CDN
service instances) for renewal, but doesn't renew anything. It is
intended to test whether you get a 'record not found' error when trying
to renew a certificate for a route that does not have a `user_data_id`
field.