Converge HTTP and DNS challenges
The broker is no longer able to function in only HTTP01 mode or DNS01
mode, both challenges have to be used, depending on the situation.

Whenever the broker is createed OR (updated AND the domains changed), we
need to do a DNS challenge, because CloudFront will reject requests made
without a host header which is known and validated with a certificate.
We can enforce this because this is user-in-the-loop action.

Whenever the broker renews a certificate, it HAS to be a HTTP challenge.
A DNS challenge will require a new DNS record, which cannot be done
automatically because the tenant will not have given us DNS access.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>