Disallow service updates
The AWS API has recently changed, requiring us to provide a field on
update that has not been provided nor required upon creation.

The field is specifically FieldLevelEncryptionID, which essentially
requires us to provide the ID of the public key, for the data to be
encrypted with.

The expectation then is that this data will be decrypted on application
side of the mix, which is not ideal scenario to be having an incident
on...

We have decided that CDN Broker have delivered enough pain to us over
the time and we're keen to deprecate it.

This doesn't come for free, and requires us to prioritise a replacement
solution to be worked on. In the mean time, we're disabling the
problematic update functionality and accept the fact we'll be handling
these as part of support responsibilities, until we have a
replacement...

Involvement: Lee Porte - TL; Mark Buckley - PM; and myself - victim.